✨ Implement category details fetching and update functionality; refactor CategoryHeader and add CollectionNavigationPanel and LinkList components for improved link management and display.
