+++ app/src/app/types.ts
@@ -1,0 +1,15 @@
+export interface Category {
+  id: string;
+  name: string;
+  is_public: boolean;
+}
+
+export interface LinkType {
+  _id: string;
+  title: string;
+  url: string;
+  visits: number;
+  category_id: string;
+  updated_id: string;
+  tags: string[];
+}
