+++ server/Endpoints/CategoryEndpoint.py
@@ -44,7 +44,6 @@ def update_category(category_id):
     new_name = data.get('name')
     # new_slug = data.get('new_slug')
     is_public = data.get('isPublic')
-    print("Received data for update:", data)
     if not category_id and not new_name:
         return jsonify({
             "success": False,
@@ -56,12 +55,10 @@ def update_category(category_id):
         new_slug = convert_to_slug(new_name)
         if category.category_slug != new_slug:
             category.category_slug= new_slug
-            category.category = new_name
+            category.name = new_name
     if is_public is not None and category.is_public != is_public:
-        print("Updating is_public from", category.is_public, "to", is_public)
         category.is_public = is_public
     category.update()
-    print(type(category), category)
     if category:
         return jsonify({
             "success": True,
@@ -74,3 +71,28 @@ def update_category(category_id):
             "message": "Category not found or you don't have permission to update it"
         }), 404
 
+@categoryRouter.route('/categories/<category_id>', methods=['GET'])
+def get_category(category_id):
+    # get token from cookies
+    token = request.cookies.get('token')
+    is_valid_token, payload = validate_and_get_token_payload(token) if token else False
+    if is_valid_token:
+        user_id = payload.get('user_id')
+    else:
+        return jsonify({
+            "success": False,
+            "message": "Invalid or missing token"
+        }), 401
+    
+    category = Category.get_by_id(category_id=category_id, user_id=user_id)
+    if category:
+        return jsonify({
+            "success": True,
+            "message": "Category details",
+            "data": category.to_json()
+        }), 200
+    else:
+        return jsonify({
+            "success": False,
+            "message": "Category not found or you don't have permission to view it"
+        }), 404
